fixed sigmoid function bug

+++ src/io.c
@@ -91,7 +91,7 @@ void writeData(Data *data) {
     
     //print: label numItem itemsInSet
     for (j = 0; j < dUserSet->numSets; j++) {
-      printf("%d %d ", (int)dUserSet->labels[j], dUserSet->uSetsSize[j]);
+      printf("%f %d ", dUserSet->labels[j], dUserSet->uSetsSize[j]);
       for (k = 0; k < dUserSet->uSetsSize[j]; k++) {
         printf("%d ", dUserSet->uSets[j][k]);
       }
